fix: full curriculum audit — Next.js 16, centavos enforcement, P0 pedagogical fixes
Three workstreams completed across 43 files:

1. Tech stack validation: Next.js 15→16 in guides/training/companion-repo,
   Zod 3→4 in public-site package.json

2. Companion repo centavos enforcement: all money displays divide by 100,
   Zod schemas enforce .int() on money fields, sanitize.ts converts
   pesos→centavos, tests updated for centavos-based output

3. Pedagogical P0 fixes (14 issues):
   - B09 capstone sends centavos to API (was sending pesos)
   - A09 optimistic update example changed from transfers to favorites
     (transfers are server-confirmed only per B03/BSP 1033)
   - A04 PinInput JSX syntax error (comment inside .map return)
   - A07 added z.coerce.number() explanation (HTML inputs return strings)
   - A21 CSP script-src now includes nonce for Next.js inline scripts
   - Bridge notes in A01/A04/A05 made self-contained (centavos explained
     inline instead of vague cross-reference to B01 Phase 6)
   - Level 5 slides token lifetime corrected to 30 min (was 15)
   - Level 6 quiz/answer key: React Compiler is separate build tool
   - Level 7 answer key: Permissions-Policy geolocation=() syntax fix
   - Level 8 exercises: clarified centavos amounts
   - Companion repo README: added missing guide references
   - B04 bridge note explaining ProtectedRoute pattern change from B02
   - A09 bridge note pointing to B04 for production auth store

+++ companion-repo/portal/src/features/accounts/api/accounts-api.ts
@@ -8,7 +8,7 @@ const accountSchema = z.object({
   name: z.string(),
   number: z.string(),
   type: z.enum(['savings', 'checking', 'time-deposit']),
-  balance: z.number(),
+  balance: z.number().int(), // centavos
   currency: z.string().default('PHP'),
   isActive: z.boolean(),
 });
