+++ companion-repo/portal/src/features/payments/api/payment-api.ts
@@ -33,9 +33,9 @@ const paymentReceiptSchema = z.object({
   billerId: z.string(),
   billerName: z.string(),
   accountId: z.string(),
-  amount: z.number(),
-  fee: z.number(),
-  total: z.number(),
+  amount: z.number().int(), // centavos
+  fee: z.number().int(), // centavos
+  total: z.number().int(), // centavos
   status: z.enum(['completed', 'pending', 'failed']),
   paidAt: z.string(),
 });
