+++ companion-repo/portal/src/schemas/account.ts
@@ -6,8 +6,8 @@ export const accountSchema = z.object({
   accountNumber: z.string().regex(/^\d{10}$/, 'Account number must be 10 digits'),
   accountName: z.string().min(1).max(100),
   type: z.enum(['savings', 'checking', 'time-deposit', 'current']),
-  balance: z.number().nonnegative(),
-  availableBalance: z.number().nonnegative(),
+  balance: z.number().int().nonnegative(), // centavos (integer)
+  availableBalance: z.number().int().nonnegative(), // centavos (integer)
   currency: z.enum(['PHP', 'USD', 'EUR', 'JPY', 'CNY']),
   status: z.enum(['active', 'dormant', 'frozen', 'closed']),
   openedDate: z.string().datetime(),
